General Refactor and rewrite of some functions

+++ src/Stocki.Domain/Models/StockQuote.cs
@@ -3,21 +3,21 @@ namespace Stocki.Domain.Models;
 public record StockQuote
 {
     public string Ticker = string.Empty;
-    public double CurrentPrice;
-    public double OpeningPrice;
-    public double ClosingPrice;
-    public double High;
-    public double Low;
+    public decimal CurrentPrice;
+    public decimal OpeningPrice;
+    public decimal ClosingPrice;
+    public decimal High;
+    public decimal Low;
 
     public StockQuote() { }
 
     public StockQuote(
         string ticker,
-        double currPrice,
-        double openingPrice,
-        double closingPrice,
-        double high,
-        double low
+        decimal currPrice,
+        decimal openingPrice,
+        decimal closingPrice,
+        decimal high,
+        decimal low
     )
     {
         Ticker = ticker;
